Better accent Color toggle UX logic
As suggested by "AshishMBakshi":
By default, the "accent color" toggle says "Use my Windows accent color" and is in the "off" position. That suggests the app is not using the system accent color, but it also does not allow the user to create a custom color, which is confusing.

When the toggle is moved to the "on" position, it then reads "Use a custom accent color," which makes sense but is inconsistent with user expectations for the behavior of a toggle in Windows apps.

A clearer alternative would be to just have the toggle always read "Use my Windows accent color" and simply have it be "on" by default.